AI-based-studyplanner-main/fie study planner/requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
requests==2.31.0
nltk==3.8.1